Improved corpus importing
